+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # CHANGELOG
 ## 1.0.2-SNAPSHOT
  * fix 多ApplicationRunner后，无法自行启动
+ * fix application run error 以及扫描不到